Fixed several compilation errors with `thread.c`

+++ user/thread.h
@@ -1,5 +1,13 @@
+#include "kernel/types.h"
+#include "kernel/riscv.h"  // Defines PGSIZE
+#include "kernel/param.h"
+#include "kernel/memlayout.h"
+#include "kernel/spinlock.h"
+#include "kernel/proc.h"
+#include "kernel/defs.h"
+
 struct lock_t {
-    uint locked;
+    uint64 locked;
 };
 
 int thread_create(void *(start_routine)(void *), void *arg);
